server/requirements.txt
python-dotenv~=1.1.1
redis~=6.4.0
motor
starlette
uvicorn
Flask~=3.1.2
requests
cloudinary~=1.44.1
PyJWT~=2.10.1
authlib~=1.6.3
werkzeug~=3.1.3
mongoengine~=0.29.1
google-auth~=2.41.1
flask-cors~=6.0.1
flask-socketio~=5.5.1
eventlet~=0.40.3
websocket-client~=1.9.0
pytest~=8.4.2
psycopg2~=2.9.11
Flask-SQLAlchemy~=3.1.1